+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change history for ui-inventory
 
+
+## [2.0.2](https://github.com/folio-org/ui-inventory/tree/v2.0.2) (2020-04-09)
+[Full Changelog](https://github.com/folio-org/ui-inventory/compare/v2.0.1...v2.0.2)
+
+* Register `instanceFormatIds` filter. Fixes UIIN-1057.
+
 ## [2.0.1](https://github.com/folio-org/ui-inventory/tree/v2.0.1) (2020-04-02)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v2.0.0...v2.0.1)